aws-ce-grafana-backend: rename aws.py to query.py
